domain-detection/requirements.txt
grpcio
grpcio-tools
transformers
torch